moving GetRolesFromUser to a utility package
